crop/requirements.txt
flask
pillow
schedule
requests